♻️ Refactor: 메인 하위 컴포넌트 폴더명 정리
