+++ src/pages/Main/components/FeedList/FeedList.styles.ts
