+++ src/pages/Main/components/HomeBox/HomeBox.styles.ts
