+++ src/pages/Main/components/NewQuote/NewQuote.styles.ts
