+++ src/pages/Main/components/NewQuote/NewQuote.tsx
