+++ src/pages/Main/components/WriteBox/WriteBox.styles.ts
