+++ src/pages/Main/components/WriteBox/WriteBox.tsx
